move KeyAuthorization to types
It's more of a data class and I think fits better alongside the
challenge types (vs the challenge handles).
