+++ src/types.rs
@@ -18,8 +18,8 @@ use x509_parser::extensions::ParsedExtension;
 #[cfg(feature = "x509-parser")]
 use x509_parser::parse_x509_certificate;
 
-use crate::BytesResponse;
 use crate::crypto::{self, KeyPair};
+use crate::{BytesResponse, Key};
 
 /// Error type for instant-acme
 #[derive(Debug, Error)]
@@ -956,6 +956,48 @@ pub(crate) enum SigningAlgorithm {
     Es256,
     /// HMAC with SHA-256,
     Hs256,
+}
+
+/// The response value to use for challenge responses
+///
+/// Refer to the methods below to see which encoding to use for your challenge type.
+///
+/// <https://datatracker.ietf.org/doc/html/rfc8555#section-8.1>
+pub struct KeyAuthorization(String);
+
+impl KeyAuthorization {
+    pub(crate) fn new(token: &str, key: &Key) -> Self {
+        Self(format!("{token}.{}", &key.thumb))
+    }
+
+    /// Get the key authorization value
+    ///
+    /// This can be used for HTTP-01 challenge responses.
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+
+    /// Get the SHA-256 digest of the key authorization
+    ///
+    /// This can be used for TLS-ALPN-01 challenge responses.
+    ///
+    /// <https://datatracker.ietf.org/doc/html/rfc8737#section-3>
+    pub fn digest(&self) -> impl AsRef<[u8]> {
+        crypto::digest(&crypto::SHA256, self.0.as_bytes())
+    }
+
+    /// Get the base64-encoded SHA256 digest of the key authorization
+    ///
+    /// This can be used for DNS-01 challenge responses.
+    pub fn dns_value(&self) -> String {
+        BASE64_URL_SAFE_NO_PAD.encode(self.digest())
+    }
+}
+
+impl fmt::Debug for KeyAuthorization {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_tuple("KeyAuthorization").finish()
+    }
 }
 
 /// Attestation payload used for device-attest-01
